+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/clientpackets/RequestCrystallizeItem.java
@@ -216,7 +216,7 @@ public class RequestCrystallizeItem extends ClientBasePacket
 				"Crystalize", _objectId, _count, activeChar, null);
 
 		// add crystals
-		int crystalId = itemToRemove.getItem().getCrystalItemId();
+		short crystalId = itemToRemove.getItem().getCrystalItemId();
 		int crystalAmount = itemToRemove.getCrystalCount();
 		L2ItemInstance createditem = activeChar.getInventory().addItem(
 				"Crystalize", crystalId, crystalAmount, activeChar,
