+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/admincommandhandlers/AdminCreateItem.java
@@ -62,7 +62,7 @@ public class AdminCreateItem implements IAdminCommandHandler {
 				if (st.countTokens()== 2)
 				{
 					String id = st.nextToken();
-					int idval = Integer.parseInt(id);
+					short idval = Short.parseShort(id);
 					String num = st.nextToken();
 					int numval = Integer.parseInt(num);
 					createItem(activeChar,idval,numval);
@@ -70,7 +70,7 @@ public class AdminCreateItem implements IAdminCommandHandler {
 				else if (st.countTokens()== 1)
 				{
 					String id = st.nextToken();
-					int idval = Integer.parseInt(id);
+					short idval = Short.parseShort(id);
 					createItem(activeChar,idval,1);
 				}
 				else
@@ -103,7 +103,7 @@ public class AdminCreateItem implements IAdminCommandHandler {
 		return (level >= REQUIRED_LEVEL);
 	}
 	
-	private void createItem(L2PcInstance activeChar, int id, int num)
+	private void createItem(L2PcInstance activeChar, short id, int num)
 	{
 		activeChar.getInventory().addItem("Admin", id, num, activeChar, null);
