+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/itemhandlers/AdvQuestItems.java
@@ -37,15 +37,15 @@ import net.sf.l2j.gameserver.serverpackets.SystemMessage;
 
 public class AdvQuestItems implements IItemHandler
 {
-    private static int[] _itemIds = { 5944, 5955, 5966, 5967, 5968, 5969, 6007, 6008, 6009, 6010 };
+    private static short[] _itemIds = { 5944, 5955, 5966, 5967, 5968, 5969, 6007, 6008, 6009, 6010 };
     
-    private static int[] award_parch = { 5922, 5923, 5924, 5925, 5926, 5927, 5928, 5929, 5930, 5931, 5932, 5933, 5934, 5935, 5936, 5937, 5938, 5939, 5940, 5941, 5942, 5943};
-    private static int[] award_gbook = { 5942, 5943, 5945, 5946, 5947, 5948, 5949, 5950, 5951, 5952, 5953, 5954};
-    private static int[] award_papy1 = { 5970, 5971, 5977, 5978, 5979, 5986, 5993, 5994, 5995, 5997, 5983, 6001};
-    private static int[] award_papy2 = { 5970, 5971, 5975, 5976, 5980, 5985, 5993, 5994, 5995, 5997, 5983, 6001};
-    private static int[] award_papy3 = { 5973, 5974, 5981, 5984, 5989, 5990, 5991, 5992, 5996, 5998, 5999, 6000, 5988, 5983, 6001};
-    private static int[] award_papy4 = { 5970, 5971, 5982, 5987, 5989, 5990, 5991, 5992, 5996, 5998, 5999, 6000, 5972, 6001};
-    private static int[] award_pouch = { 6011, 6012, 6013, 6014, 6015, 6016, 6018, 6019, 6020};
+    private static short[] award_parch = { 5922, 5923, 5924, 5925, 5926, 5927, 5928, 5929, 5930, 5931, 5932, 5933, 5934, 5935, 5936, 5937, 5938, 5939, 5940, 5941, 5942, 5943};
+    private static short[] award_gbook = { 5942, 5943, 5945, 5946, 5947, 5948, 5949, 5950, 5951, 5952, 5953, 5954};
+    private static short[] award_papy1 = { 5970, 5971, 5977, 5978, 5979, 5986, 5993, 5994, 5995, 5997, 5983, 6001};
+    private static short[] award_papy2 = { 5970, 5971, 5975, 5976, 5980, 5985, 5993, 5994, 5995, 5997, 5983, 6001};
+    private static short[] award_papy3 = { 5973, 5974, 5981, 5984, 5989, 5990, 5991, 5992, 5996, 5998, 5999, 6000, 5988, 5983, 6001};
+    private static short[] award_papy4 = { 5970, 5971, 5982, 5987, 5989, 5990, 5991, 5992, 5996, 5998, 5999, 6000, 5972, 6001};
+    private static short[] award_pouch = { 6011, 6012, 6013, 6014, 6015, 6016, 6018, 6019, 6020};
     
     private final static Random _rnd = new Random();
     
@@ -53,8 +53,8 @@ public class AdvQuestItems implements IItemHandler
     {
         if (!(playable instanceof L2PcInstance)) return;
         L2PcInstance activeChar = (L2PcInstance)playable;
-        int itemId = item.getItemId();
-        int itemToCreateId = itemId;
+        short itemId = item.getItemId();
+        short itemToCreateId = itemId;
 
         switch (itemId)
         {
@@ -95,7 +95,7 @@ public class AdvQuestItems implements IItemHandler
         ItemList playerUI = new ItemList(activeChar, false);
         activeChar.sendPacket(playerUI);
     }
-    public int[] getItemIds()
+    public short[] getItemIds()
     {
         return _itemIds;
     }
