+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/itemhandlers/MercTicket.java
@@ -84,7 +84,7 @@ public class MercTicket implements IItemHandler
     }
 
     // left in here for backward compatibility
-    public int[] getItemIds()
+    public short[] getItemIds()
     {
         return MercTicketManager.getInstance().getItemIds();
     }
